Create service module to handle todo commands and implement addTodo function
